okhttp3 socket connect e2e within known dns lookup
