--- e2e-tests/spring-rest-openapi-v3/src/main/kotlin/com/foo/rest/examples/spring/openapi/v3/wiremock/socketconnect/WmSocketConnectDto.kt
@@ -1,6 +1,0 @@
-package com.foo.rest.examples.spring.openapi.v3.wiremock.socketconnect
-
-class WmSocketConnectDto {
-
-    var x: Int? = null
-}
