+++ e2e-tests/spring-rest-openapi-v3/src/main/kotlin/com/foo/rest/examples/spring/openapi/v3/wiremock/socketconnect/okhttp3/WmSocketConnectDto.kt
@@ -1,0 +1,6 @@
+package com.foo.rest.examples.spring.openapi.v3.wiremock.socketconnect.okhttp3
+
+class WmSocketConnectDto {
+
+    var x: Int? = null
+}
