+++ e2e-tests/spring-rest-openapi-v3/src/test/kotlin/com/foo/rest/examples/spring/openapi/v3/wiremock/socketconnect/okhttp3/WmSocketConnectController.kt
@@ -1,4 +1,4 @@
-package com.foo.rest.examples.spring.openapi.v3.wiremock.socketconnect
+package com.foo.rest.examples.spring.openapi.v3.wiremock.socketconnect.okhttp3
 
 import com.foo.rest.examples.spring.openapi.v3.SpringController
